Fix Bis hand refill bug.
Closes #1

Until now, the game was saved in some actions of the model. That lead to
the following behaviour:
When a player plays on a bis, at the end of the turn the player's hand
was refilled, but the method game.set_next_player was not called, and
therefore, no game.save() was called, leading to the player's new hand
to not be saved, which duplicated the tokens they just played.

The solution for that is to explicitly say in the service when the game
has to be saved, rather than relying on the model to save itself in the
good times, or not.

The bug comes from the transition from games saved in memory (in which
case any manipulation on the game was naturally saving it) to the games
saves in DB (in which case an explicit SQL INSERT/UPDATE had to be
called).

Another side effect of the previous state of the code was that multiple
actions on a game would lead to multiple SQL updates, where a single one
would be needed.